try to use the dieselcrate in alpine
looks like it's trying to link with /lib/ld64.so.1
which looks wrong.

i tried hacking it so that it would use musl's ld instead, but this
causes a terrible stack trace from an unwrap in `lang_start`

```sh
  12: __rust_maybe_catch_panic
             at ./checkout/src/libpanic_unwind/lib.rs:99
  13: std::rt::lang_start
             at ./checkout/src/libstd/panicking.rs:459
             at ./checkout/src/libstd/panic.rs:361
             at ./checkout/src/libstd/rt.rs:59
```

the way to go is probably to get the image to use the correct ld during
cargo build.