Install the minimal rustup profile
Rustup 1.12.0 introduced profiles and changed the default installation
to include more components. Now, rustfmt and clippy are included by
default in the "default" profile. A new "minimal" profile is also
available and should be used in CI from now on. Compared to the previous
default install, it is lighter as it does not include rust-docs.

The difference in default components can be seen in recent CI runs, one
before and one after the new rustup release:
2019-10-11: https://travis-ci.org/clux/muslrust/builds/596639602#L797-L800
2019-10-15: https://travis-ci.org/clux/muslrust/builds/598191165#L798-L803

[Rustup 1.12.0]: https://blog.rust-lang.org/2019/10/15/Rustup-1.20.0.html